refactor: extract chat UI into chat-ui.js
Created chat-ui.js with a single exported initChatUI(CURRENT_USER) function containing all chat UI logic - channel switching, DM management, message rendering, online users list, send message, open/close handlers
main.js DOMContentLoaded now just calls initChat(), createUserManager(), handleRoute() and initChatUI() - significantly leaner
Removed activeChannel and messageHistory from main.js - they now live inside chat-ui.js
Added comprehensive comments throughout chat-ui.js
