+++ frontend/src/main.js
@@ -2,9 +2,10 @@ import "./styles.css";
 import { Engine, Scene } from "@babylonjs/core";
 import { initGameScene } from "./game.js";
 import { createUserManager } from './usermanagement.js';
-import { initChat, sendChatMessage, onlineUsers, initTyping } from './chat.js';
+import { initChat } from './chat.js';
 import { getCurrentUser as fetchCurrentUser } from './usermanagement.js';
 import { initI18n, t, TranslationKey, updatePageTranslations, setLanguage, getCurrentLanguage, Language } from "./i18n";
+import { initChatUI } from './chat-ui.js';
 
 // --- Game Variables ---
 let ws = null;
@@ -476,332 +477,11 @@ window.addEventListener("DOMContentLoaded", async () => {
 	console.log("Current user:", CURRENT_USER);
 	initChat();
 
-	// DOM elements
-	const chatContainer = document.getElementById("chatContainer");
-	const openChatBtn = document.getElementById("openChatBtn");
-	const closeChatBtn = document.getElementById("closeChatBtn");
-	const sendChatBtn = document.getElementById("sendChatBtn");
-	const chatInput = document.getElementById("chatInput");
-	const onlineUsersList = document.getElementById("onlineUsersList");
-	const channelTabs = document.getElementById("channelTabs");
-	const chatMessages = document.getElementById("chatMessages");
-	const channelTitle = document.getElementById("channelTitle");
-
-	// State management
-	let activeChannel = "global"; // Current active channel (global or user ID)
-	const messageHistory = {
-		global: [] // Each channel stores its messages
-	};
-
-	// Initialize i18n (if you have it)
-	// initI18n();
-	// setLanguage(Language.EN);
-	// updatePageTranslations();
-
 	// Create user manager UI
 	createUserManager();
 
 	// Initial route handling
 	handleRoute(window.location.pathname);
 
-	// ============================================
-	// CHANNEL MANAGEMENT
-	// ============================================
-
-	function switchChannel(channelId) {
-		activeChannel = channelId;
-
-		// Update tab active state
-		document.querySelectorAll(".channel-tab").forEach(tab => {
-			tab.classList.remove("active");
-		});
-		const activeTab = document.querySelector(`[data-channel="${channelId}"]`);
-		if (activeTab) {
-			activeTab.classList.add("active");
-		}
-
-		// Update channel title
-		if (channelId === "global") {
-			channelTitle.textContent = "# Global Chat";
-		} else {
-			const user = onlineUsers.find(u => u.id === channelId);
-			channelTitle.textContent = user ? `@ ${user.name}` : "@ Direct Message";
-		}
-
-		// Display messages for this channel
-		renderMessages(channelId);
-		
-		// Focus input
-		chatInput.focus();
-	}
-
-	function openDMChannel(userId, userName, switchToChannel = true) {
-		// Don't open DM with yourself
-		if (userId === CURRENT_USER.user_id) return;
-
-		// Check if tab already exists
-		const existingTab = document.querySelector(`[data-channel="${userId}"]`);
-		if (existingTab) {
-			switchChannel(userId);
-			return;
-		}
-
-		// Create message history for this channel if it doesn't exist
-		if (!messageHistory[userId]) {
-			messageHistory[userId] = [];
-		}
-
-		// Create new tab
-		const tab = document.createElement("button");
-		tab.className = "channel-tab";
-		tab.dataset.channel = userId;
-
-		tab.innerHTML = `
-			<span class="font-bold opacity-80">@</span>
-			<span>${userName}</span>
-			<span class="close-tab" data-close="${userId}">  X</span>
-		`;
-
-		// Insert before close button
-		channelTabs.appendChild(tab);
-
-		// Add click handlers
-		tab.addEventListener("click", (e) => {
-			if (e.target.classList.contains("close-tab")) {
-				// Close tab
-				e.stopPropagation();
-				closeDMChannel(userId);
-			} else {
-				// Switch to this channel
-				switchChannel(userId);
-			}
-		});
-
-		// Only switch if requested
-		if (switchToChannel) switchChannel(userId);
-	}
-
-	function closeDMChannel(userId) {
-		// Remove tab
-		const tab = document.querySelector(`[data-channel="${userId}"]`);
-		if (tab) tab.remove();
-
-		// If this was the active channel, switch to global
-		if (activeChannel === userId) {
-			switchChannel("global");
-		}
-
-		// Optionally: keep message history in case they reopen
-		// Or delete it: delete messageHistory[userId];
-	}
-
-	// ============================================
-	// MESSAGE MANAGEMENT
-	// ============================================
-
-	function addMessage(channelId, message) {
-		// Initialize channel history if needed
-		if (!messageHistory[channelId]) {
-			messageHistory[channelId] = [];
-		}
-
-		// Add message to history
-		messageHistory[channelId].push(message);
-
-		// If this is the active channel, render it immediately
-		if (channelId === activeChannel) {
-			renderMessages(channelId);
-		} else {
-			// Show unread badge on tab
-			const tab = document.querySelector(`[data-channel="${channelId}"]`);
-			if (tab && !tab.querySelector(".unread-badge")) {
-				const badge = document.createElement("span");
-				badge.className = "unread-badge";
-				badge.textContent = "1";
-				tab.appendChild(badge);
-			} else if (tab) {
-				const badge = tab.querySelector(".unread-badge");
-				badge.textContent = parseInt(badge.textContent) + 1;
-			}
-		}
-	}
-
-	function renderMessages(channelId) {
-		chatMessages.innerHTML = "";
-
-		const messages = messageHistory[channelId] || [];
-		messages.forEach(msg => {
-			const msgDiv = document.createElement("div");
-			msgDiv.className = "chat-message text-base leading-relaxed text-gray-200";
-
-			const isOwnMessage = msg.senderId === CURRENT_USER.user_id;
-			if (isOwnMessage) {
-				msgDiv.classList.add("self");
-			} else if (channelId !== "global") {
-				msgDiv.classList.add("dm-received");
-			}
-
-			const senderSpan = document.createElement("span");
-			senderSpan.className = "sender";
-			senderSpan.textContent = isOwnMessage ? "Me" : msg.senderName;
-
-			const messageSpan = document.createElement("span");
-			messageSpan.textContent = msg.message;
-
-			msgDiv.appendChild(senderSpan);
-			msgDiv.appendChild(document.createTextNode(": "));
-			msgDiv.appendChild(messageSpan);
-
-			chatMessages.appendChild(msgDiv);
-		});
-
-		// Scroll to bottom
-		chatMessages.scrollTop = chatMessages.scrollHeight;
-
-		// Clear unread badge if switching to this channel
-		const tab = document.querySelector(`[data-channel="${channelId}"]`);
-		if (tab) {
-			const badge = tab.querySelector(".unread-badge");
-			if (badge) badge.remove();
-		}
-	}
-
-	// ============================================
-	// USER LIST
-	// ============================================
-
-	function renderOnlineUsers() {
-	if (!onlineUsersList) return;
-	
-	onlineUsersList.innerHTML = "";
-
-	if (!onlineUsers || onlineUsers.length === 0) {
-		onlineUsersList.innerHTML = '<div class="text-xs text-gray-500 px-2">No users online</div>';
-		return;
-	}
-
-	onlineUsers.forEach(user => {
-		// Skip yourself — every user past this point is guaranteed to be someone else
-		if (user.id === CURRENT_USER.user_id) return;
-
-		const div = document.createElement("div");
-		div.className = "user-item";
-
-		// Online status dot
-		const statusDot = document.createElement("span");
-		statusDot.className = "w-2 h-2 rounded-full bg-[#00FF00] flex-shrink-0";
-
-		// Username
-		const nameSpan = document.createElement("span");
-		nameSpan.textContent = user.name || user.id;
-
-		div.appendChild(statusDot);
-		div.appendChild(nameSpan);
-
-		// Click to open DM
-		div.addEventListener("click", () => {
-			openDMChannel(user.id, user.name || user.id);
-		});
-
-		onlineUsersList.appendChild(div);
-	});
-}
-
-	// Listen for online users updates
-	window.addEventListener("onlineUsersUpdated", renderOnlineUsers);
-
-	// Listen for incoming messages (from chat.js)
-	window.addEventListener("chatMessageReceived", (e) => {
-		const { channelId, message } = e.detail;
-		
-		console.log("chatMessageReceived event fired!");
-		console.log("channelId:", channelId);
-		console.log("message:", message);
-		console.log("activeChannel:", activeChannel);
-		
-		// If this is a DM channel and the tab doesn't exist, create it
-		if (channelId !== "global") {
-			const existingTab = document.querySelector(`[data-channel="${channelId}"]`);
-			if (!existingTab) {
-				// Create DM channel but DON'T switch to it
-				openDMChannel(channelId, message.senderName, false)
-			}
-		}
-		// Just add the message - don't auto-switch channels
-		addMessage(channelId, message);
-	});
-
-	// ============================================
-	// CHAT OPEN/CLOSE
-	// ============================================
-
-	if (openChatBtn && chatContainer) {
-		openChatBtn.style.display = "block";
-		openChatBtn.addEventListener("click", () => {
-			chatContainer.style.display = "flex";
-			openChatBtn.style.display = "none";
-			chatInput.focus();
-			renderOnlineUsers();
-		});
-	}
-
-	if (closeChatBtn && chatContainer && openChatBtn) {
-		closeChatBtn.addEventListener("click", () => {
-			chatContainer.style.display = "none";
-			openChatBtn.style.display = "block";
-		});
-	}
-
-	// ============================================
-	// SEND MESSAGE
-	// ============================================
-
-	initTyping(chatInput);
-
-	if (sendChatBtn && chatInput) {
-		const sendMessage = () => {
-			const message = chatInput.value.trim();
-			if (!message) return;
-
-			// Determine target
-			const target = activeChannel === "global" ? null : activeChannel;
-
-			// Send via WebSocket
-			sendChatMessage(message, target);
-
-			chatInput.value = "";
-			chatInput.focus();
-		};
-
-		sendChatBtn.addEventListener("click", sendMessage);
-		
-		chatInput.addEventListener("keypress", e => {
-			if (e.key === "Enter" && !e.shiftKey) {
-				e.preventDefault();
-				sendMessage();
-			}
-		});
-	}
-
-	// ============================================
-	// GLOBAL CHANNEL TAB CLICK
-	// ============================================
-	
-	console.log("Looking for global tab...");
-	const globalTab = document.querySelector('[data-channel="global"]');
-	console.log("Global tab element found:", globalTab)
-	
-	const allTabs = document.querySelectorAll('.channel-tab');
-	console.log("All channel tabs:", allTabs);
-	allTabs.forEach(tab => {
-		console.log("Tab data-channel:", tab.dataset.channel);
-	});
-	
-	if (globalTab) {
-		globalTab.addEventListener("click", () => {
-			console.log("Global tab clicked!");
-			switchChannel("global");
-		});
-	}
-
+	initChatUI(CURRENT_USER);
 });
